Move use_event_ref/1 to Filament.Experimental.Hooks
Not ready to commit to a stable API yet; users opt in with
`import Filament.Experimental.Hooks`.
